Separate shop & shopItem handler and router files
